Refactor premium tests by splitting utils to own file
